requirements.txt
# Core dependencies
sentence-transformers>=2.2.0
chromadb>=0.4.0
mistralai>=1.0.0
transformers>=4.30.0
torch>=2.0.0

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Document processing
pypdf>=3.0.0
python-docx>=1.0.0

# Utilities
numpy>=1.24.0
